requirements.txt
openai==1.77.0
pillow==11.2.1
tqdm==4.67.1
